+++ CHANGES.rst
@@ -17,9 +17,12 @@ Major Changes
 Minor Changes
 ^^^^^^^^^^^^^
 
+- Remove ``asynchs.loops.stop``, add ``asynchs.loops.set_event_loop``.
+- Improvements & fixes to type hints.
 - Improvements & fixes to doc strings.
 - Improvements & fixes to package tests.
 - Improvements & fixes to package documentation.
+- Improvements & fixes to code formatting & style.